Introduce Provisioner interface
Controller's CreateVolume and DeleteVolume now call method of Provisioner interface. All heavy-lifting logic of create/delete specific for access-point provsioning is hidden now in AccessPointProvisioner struct.
